Add int for multiplicity

+++ pydefect/input_maker/supercell_maker.py
@@ -173,7 +173,8 @@ class Supercells:
 
         for i in range(int(max_num_atoms / len(primitive_cell))):
             isotropy, angle = calc_isotropy(self.unitcell, trans_mat)
-            multiplicity = round(np.linalg.det(trans_mat))
+            # int is needed when numpy.float is rounded.
+            multiplicity = int(round(np.linalg.det(trans_mat)))
             num_atoms = multiplicity * len(primitive_cell)
             if num_atoms > max_num_atoms:
                 break
